src/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
requests==2.32.3
httpx==0.27.2
beautifulsoup4==4.12.3
gunicorn==23.0.0



